requirements.txt
django==4.2.11
djangorestframework==3.14.0
drf-spectacular===0.27.1
whitenoise===6.6.0